requirements.txt
numpy
scipy
tqdm
matplotlib

open3d
opencv-python

torch-scatter

jupyter
imageio

glog
icecream

pandas
pyrealsense2

pybind11
gdown

colored_glog

open-gopro==0.12.0
paho-mqtt==1.6.1